+++ src/components/ItemEditCard.vue
@@ -2,10 +2,11 @@
   <div class="card block">
     <div class="card-content">
       <div class="media">
-        <div class="media-content">
+        <div class="media-content" @click="linkEdit">
           <!-- <h4 class="bold">Kushikatsu Special Platter</h4> -->
           <p class="item-name">
-            {{ menuitem.itemName }}
+              {{ menuitem.itemName }}
+            </router-link>
           </p>
           <p class="item-price">
             {{ $n(menuitem.price, "currency") }}
@@ -65,6 +66,11 @@ export default {
     };
   },
   methods: {
+    linkEdit() {
+      this.$router.push({
+        path: `/admin/restaurants/${this.restaurantId()}/menus/${this.menuitem.id}`
+      });
+    },
     positionUp() {
       this.$emit("positionUp", this.menuitem.id);
     },
